LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/track/track-element/track-cues-pause-on-exit-expected.txt



PASS Video is paused after cues having pause-on-exit flag are processed
